+++ src/codegen/mod.rs
@@ -1,6 +1,7 @@
 mod interference_graph;
 mod translator;
 mod ssa_elimination;
+mod backpatch;
 
 #[cfg(test)]
 mod tests;
@@ -11,6 +12,7 @@ pub use interference_graph::{InterferenceGraph, find_copy_edges};
 pub use crate::runtime::vm::{ByteCode, Func};
 use translator::ByteCodeTranslator;
 use ssa_elimination::ssa_elimination;
+use backpatch::{BackPatchLabel, push_jump_instr, push_generic_jump_instr, back_patch_jump_instructions};
 
 use crate::ir::{Block, Func as IRFunc, LabelID, Tac};
 
@@ -126,57 +128,4 @@ fn generate_bytecode(ir_func: &IRFunc, graph: &InterferenceGraph, max_clique: us
     func
 }
 
-#[derive(Eq, PartialEq, Hash, Copy, Clone)]
-enum BackPatchLabel {
-    Label(LabelID),
-    Temp(usize),
-}
-
-fn push_jump_instr(func: &mut Func, jump_positions: &mut HashMap<BackPatchLabel, Vec<usize>>, label: BackPatchLabel) {
-    let instr = ByteCode::Jump { offset: 0 };
-
-    push_generic_jump_instr(instr, func, jump_positions, label);
-}
-
-fn push_generic_jump_instr(instr: ByteCode, func: &mut Func, jump_positions: &mut HashMap<BackPatchLabel, Vec<usize>>, label: BackPatchLabel) {
-    let position = func.len();
-
-    func.push_instr(instr);
-
-    if let Some(positions) = jump_positions.get_mut(&label) {
-        positions.push(position);
-    } else {
-        jump_positions.insert(label, vec![position]);
-    }
-}
-
-fn back_patch_jump_instructions(
-    func: &mut Func, 
-    label_positions: HashMap<BackPatchLabel, usize>,
-    jump_positions: HashMap<BackPatchLabel, Vec<usize>>,
-) {
-    let instrs = func.get_instrs_mut();
-    for (label, positions) in jump_positions.iter() {
-        let label_position = label_positions.get(label).unwrap();
-
-        for jump_pos in positions.iter() {
-            match &mut instrs[*jump_pos] {
-                ByteCode::Jnt { offset, .. } |
-                ByteCode::Jit { offset, .. } |
-                ByteCode::Jump { offset } => {
-                    let abs_diff: usize = label_position.abs_diff(*jump_pos);
-                    let signed_offset: isize =
-                    if *label_position < *jump_pos {
-                        isize::try_from(abs_diff).unwrap() * -1
-                    } else {
-                        isize::try_from(abs_diff).unwrap()
-                    };
-
-                    *offset = i16::try_from(signed_offset).unwrap();
-                }
-                _ => panic!("CODEGEN ERROR during Back Patching")
-            }
-        }
-    }
-}
